Shader now integrated with game, replacing all functionality of the former tile-map.

+++ Assets/Scripts/GameState.cs
@@ -11,7 +11,7 @@ public class GameState : MonoBehaviour
     public List<Cohort> activeCohorts = new List<Cohort>();
 
     //NOTE: CPU becomes a limitiation somewhere between one and ten million tiles on-screen. Memory usage is also significant.
-    public short [,] map;
+    public int [,] map;
     public int mapOffset;
 
     ViewManager vManage;
@@ -19,7 +19,7 @@ public class GameState : MonoBehaviour
     void Awake () {
 //this is in Awake rather than Start so that the array gets made before other scripts try to access it.
         int mapSize = GetComponent<setup>().mapSize;
-        map = new short [mapSize,mapSize];
+        map = new int [mapSize,mapSize];
         mapOffset = map.GetLength(0) / 2;
         vManage = GameObject.Find("Player Perspective").GetComponent<ViewManager>();
     }
